requirements.txt
streamlit
streamlit-pdf-viewer
PyMuPDF # fitz
pdfminer.six==20250324 # Explicitly specify latest available version
pdfplumber
pytesseract
nougat-ocr==0.1.17
albumentations==1.4.18 # Use latest available version for Python 3.8
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
# grobid-client-python # Placeholder - Check GROBID client documentation
langchain
langchain-community # Often needed for specific LangChain components
tenacity==8.3.0 # Pin to compatible version for langchain
python-docx
